Fixes #1. Update sample dewpoint calculator app.
The dewpoint calculator app was crashing with this error:

  qwaylandglcontext.cpp:386: QWaylandGLContext::makeCurrent: eglError: 3009, this: 0x8e97d08

This commit fixes the crash by using the qml.SailfishRun() methods to
launch the SailfishApp:: instance, per this thread:

 https://together.jolla.com/question/105098/how-to-setup-go-142-runtime-and-go-qml-pkg-for-mersdk/

Also found that X-Nemo-Application-Type=generic did not work and never
launced the app. Instead setting this to
X-Nemo-Application-Type=no-invoker fixed the problem.

Tested this using SailfishOSSDK-Beta-1511-Qt5-linux-64 and is working
using the MerSDK and Jolla Emu.